My 3rd commit after writing some scenarios and definitions
