--- src/test/java/com/fidexio/step_definitions/FIDEXI_547_Employee_StepDefinitions.java
@@ -1,55 +1,0 @@
-package com.fidexio.step_definitions;
-
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-
-public class FIDEXI_547_Employee_StepDefinitions {
-
-    @Given("Pos Manager logged in to the homepage and clicked on the Employees module")
-    public void pos_manager_logged_in_to_the_homepage_and_clicked_on_the_employees_module() {
-
-    }
-
-    @When("Pos Manager click on the Create button")
-    public void pos_manager_click_on_the_create_button() {
-
-    }
-
-    @Then("Pos Manager should be able to land on the new employee information page")
-    public void pos_manager_should_be_able_to_land_on_the_new_employee_information_page() {
-
-    }
-
-    @When("Pos Manager click on the Import button")
-    public void pos_manager_click_on_the_ımport_button() {
-
-    }
-
-    @Then("Pos Manager should be able to land on the Import a File page")
-    public void pos_manager_should_be_able_to_land_on_the_ımport_a_file_page() {
-
-    }
-
-    @When("Pos Manager click on the Left Arrow - \\(<) button")
-    public void pos_manager_click_on_the_left_arrow_button() {
-
-    }
-
-    @Then("PosManager should be able to click on the Left Arrow - \\(<) button.")
-    public void pos_manager_should_be_able_to_click_on_the_left_arrow_button() {
-
-    }
-
-    @When("Pos Manager click on the Right Arrow - \\(>) button.")
-    public void pos_manager_click_on_the_right_arrow_button() {
-
-    }
-
-    @Then("Pos Manager should be able to click on the Right Arrow - \\(>) button.")
-    public void pos_manager_should_be_able_to_click_on_the_right_arrow_button() {
-
-    }
-
-
-}
