+++ src/test/java/com/fidexio/step_definitions/US_77_FIDEXI_547_Employees_StepDefinitions.java
@@ -1,0 +1,90 @@
+package com.fidexio.step_definitions;
+
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class US_77_FIDEXI_547_Employees_StepDefinitions {
+
+    @Given("Pos Manager logged in to the homepage and clicked on the Employees module")
+    public void pos_manager_logged_in_to_the_homepage_and_clicked_on_the_employees_module() {
+
+    }
+
+    @When("Pos Manager click on the Create button")
+    public void pos_manager_click_on_the_create_button() {
+
+    }
+
+    @Then("Pos Manager should be able to land on the new employee information page")
+    public void pos_manager_should_be_able_to_land_on_the_new_employee_information_page() {
+
+    }
+
+    @When("Pos Manager click on the import button")
+    public void pos_manager_click_on_the_import_button() {
+
+    }
+
+    @Then("Pos Manager should be able to land on the import a File page")
+    public void pos_manager_should_be_able_to_land_on_the_import_a_file_page() {
+
+    }
+
+    @When("Pos Manager click on the Left Arrow - \\(<) button")
+    public void pos_manager_click_on_the_left_arrow_button() {
+
+    }
+
+    @Then("PosManager should be able to click on the Left Arrow - \\(<) button.")
+    public void pos_manager_should_be_able_to_click_on_the_left_arrow_button() {
+
+    }
+
+    @When("Pos Manager click on the Right Arrow - \\(>) button.")
+    public void pos_manager_click_on_the_right_arrow_button() {
+
+    }
+
+    @Then("Pos Manager should be able to click on the Right Arrow - \\(>) button.")
+    public void pos_manager_should_be_able_to_click_on_the_right_arrow_button() {
+
+    }
+
+    @When("Pos Manager click on the Kanban button")
+    public void pos_manager_click_on_the_kanban_button() {
+
+    }
+
+    @Then("Pos Manager should be able to click on the Kanban button and should see the list of items in the Kanban style")
+    public void pos_manager_should_be_able_to_click_on_the_kanban_button_and_should_see_the_list_of_items_in_the_kanban_style() {
+
+    }
+
+    @When("Pos Manager click on the List button")
+    public void pos_manager_click_on_the_list_button() {
+
+    }
+
+    @Then("Pos Manager should be able to click on the List button and should see the list of items in the List style")
+    public void pos_manager_should_be_able_to_click_on_the_list_button_and_should_see_the_list_of_items_in_the_list_style() {
+
+    }
+
+    @When("Pos Manager enters employee credentials")
+    public void pos_manager_enters_employee_credentials() {
+
+    }
+
+    @When("Pos Manager click on the Save button")
+    public void pos_manager_click_on_the_save_button() {
+
+    }
+
+    @Then("PosManager should be able to click on the Save button and should be able to create a new Employee id with the Employee created message")
+    public void pos_manager_should_be_able_to_click_on_the_save_button_and_should_be_able_to_create_a_new_employee_id_with_the_employee_created_message() {
+
+    }
+
+
+}
